backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Каманин Юрий <kamanchi22@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sorl-thumbnail = "^12.9.0"
djoser = "^2.2.0"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
